+++ NBitcoin.Altcoins/Groestlcoin.Internals/DigestEngine.cs
@@ -158,7 +158,7 @@ internal abstract class DigestEngine : Digest {
   /** @see Digest */
   public void update(byte input)
   {
-    inputBuf[inputLen ++] = (byte)input;
+    inputBuf[inputLen ++] = input;
     if (inputLen == blockLen) {
       processBlock(inputBuf);
       blockCount ++;
